🐞 fix(type def): from usize to i32
As proto3 type system has not 'usize' type, I cannot pass a Hashmap with a usize type...

+++ tra/src/replica/file_watcher.rs
@@ -17,7 +17,7 @@ lazy_static! {
 pub struct FileWatcher {
     pub inotify: Inotify,
     pub wd_map: Arc<RwLock<HashMap<WatchDescriptor, PathBuf>>>,
-    pub cnt: usize,
+    pub cnt: i32,
 }
 
 #[derive(Clone)]
